requirements.txt
asyncpg==0.30.0
pydantic==2.9.2
python-dotenv==1.0.1
pytest==9.0.2